Passing the management url from the server
