+++ src/stripe/stripe-service.ts
@@ -10,15 +10,13 @@ import type {
 import type { BrandingInfoResponse } from "../networking/responses/branding-response";
 import { Theme } from "../ui/theme/theme";
 import { DEFAULT_TEXT_STYLES } from "../ui/theme/text";
-import type {
-  StripeElementsConfiguration,
-  StripeExpressCheckoutConfiguration,
-} from "../networking/responses/stripe-elements";
+import type { StripeElementsConfiguration } from "../networking/responses/stripe-elements";
 import type { Product, SubscriptionOption } from "../entities/offerings";
 import type { Translator } from "../ui/localization/translator";
 import { LocalizationKeys } from "../ui/localization/supportedLanguages";
 import type { StripeExpressCheckoutElementOptions } from "@stripe/stripe-js/dist/stripe-js/elements/index";
 import { type Period, PeriodUnit } from "../helpers/duration-helper";
+import type { StripeExpressCheckoutConfiguration } from "./stripe-express-checkout-configuration";
 
 export enum StripeServiceErrorCode {
   ErrorLoadingStripe = 0,
@@ -389,6 +387,7 @@ export class StripeService {
     productDetails: Product,
     subscriptionOption: SubscriptionOption,
     translator: Translator,
+    managementUrl: string,
   ): StripeExpressCheckoutConfiguration {
     const priceMicros = subscriptionOption.base.price?.amountMicros;
 
@@ -413,7 +412,7 @@ export class StripeService {
       applePay: {
         recurringPaymentRequest: {
           paymentDescription: productDetails.title,
-          managementURL: "https://example.com/manage",
+          managementURL: managementUrl,
           trialBilling: hasTrial
             ? {
                 label: translator.translate(LocalizationKeys.ApplePayFreeTrial),
